requirements.txt
pandas==2.3.2

joblib==1.5.2

sentence_transformers==5.1.0

groq==0.31.1

python-dotenv==1.1.1

scikit-learn==1.7.2